requirements.txt
openai
python-dotenv
pydantic
jsonschema
beautifulsoup4
requests
playwright
tinycss2
flask
